transaction-checks: factor out transaction input checks into a reusable crate
